requirements.txt
fastapi
uvicorn
pyyaml
pandas